Refactor: Clean up src/shared directory based on audit
This commit performs a cleanup of the `src/shared` directory based on the recommendations in `shared_audit.md`.

The following changes were made:

- **Consolidated Demo Data**: Merged `demoModuleData.ts`, `demoSlideDeckData.ts`, and `testDemoSlideDeck.ts` into a new `src/shared/demos/` directory to centralize test and demo data. Updated all relevant import paths.

- **Refactored Type Definitions**:
    - Moved utility functions (`detectVideoSource`, `extractYouTubeVideoId`) from `types.ts` to a new, more appropriate location at `src/client/utils/videoUtils.ts`.
    - Consolidated `ViewerModes` from `interactiveTypes.ts` into `slideTypes.ts`.
    - Deleted `interactiveTypes.ts` as it became redundant after its types were moved.
    - Updated all import paths affected by these changes.

- **Removed Dead Code**:
    - Deleted the unused barrel file `src/shared/index.ts`.
    - Removed unused exported functions from `hotspotStylePresets.ts` and `migrationUtils.ts` to reduce clutter and improve code clarity.

These changes address the duplication and legacy code issues identified in the audit, improving the overall structure and maintainability of the shared codebase.
