+++ src/client/components/App.tsx
@@ -2,8 +2,8 @@ import React, { useState, useCallback, useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { AuthProvider, useAuth } from '../../lib/authContext';
 import { appScriptProxy } from '../../lib/firebaseProxy';
-import { demoModuleData } from '../../shared/demoModuleData';
-import { createDemoSlideDeck } from '../../shared/demoSlideDeckData';
+import { demoModuleData } from '../../shared/demos/module';
+import { createDemoSlideDeck } from '../../shared/demos/slideDeck';
 import { SlideDeck, ThemePreset, BackgroundMedia } from '../../shared/slideTypes';
 import { Project, InteractiveModuleState } from '../../shared/types';
 import { ensureProjectCompatibility } from '../../shared/migrationUtils';
