+++ src/client/components/testing/EditorTestPage.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useCallback } from 'react';
 import { SlideDeck, ThemePreset, ElementInteraction } from '../../../shared/slideTypes';
-import { createTestDemoSlideDeck } from '../../../shared/testDemoSlideDeck';
+import { createTestDemoSlideDeck } from '../../../shared/demos/test';
 import { InteractiveModuleState } from '../../../shared/types';
 import { Z_INDEX } from '../../utils/zIndexLevels';
 import ModernSlideEditor from '../editors/ModernSlideEditor';
