--- src/shared/index.ts
@@ -1,3 +1,0 @@
-// Barrel exports for shared modules
-export * from './types';
-export * from './slideTypes';
